auto-upload SPROCs to CosmosDB
