+++ extensions/azure/cosmos/contract-negotiation-store-cosmos/src/main/java/org/eclipse/dataspaceconnector/contract/negotiation/store/CosmosContractNegotiationStoreConfig.java
@@ -27,6 +27,16 @@ public class CosmosContractNegotiationStoreConfig extends AbstractCosmosConfig {
         super(context);
     }
 
+    /**
+     * Boolean setting to allow or disallow auto-uploading any stored procedures that this extension needs.
+     * Disable to reduce startup times.
+     *
+     * @return the key of the setting
+     */
+    public String allowSprocAutoUploadSetting() {
+        return "edc.contractnegotiationstore.cosmos.allow.sproc.autoupload";
+    }
+
     @Override
     protected String getAccountNameSetting() {
         return "edc.contractnegotiationstore.cosmos.account-name";
