+++ extensions/azure/cosmos/transfer-process-store-cosmos/src/main/java/org/eclipse/dataspaceconnector/transfer/store/cosmos/CosmosTransferProcessStoreExtension.java
@@ -25,6 +25,7 @@ import org.eclipse.dataspaceconnector.spi.system.health.HealthCheckService;
 import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
 import org.eclipse.dataspaceconnector.transfer.store.cosmos.model.TransferProcessDocument;
 
+
 /**
  * Provides an in-memory implementation of the {@link org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore} for testing.
  */
@@ -59,6 +60,10 @@ public class CosmosTransferProcessStoreExtension implements ServiceExtension {
 
         healthService.addReadinessProvider(() -> cosmosDbApi.get().forComponent(name()));
 
+        if (context.getSetting(configuration.allowSprocAutoUploadSetting(), true)) {
+            cosmosDbApi.uploadStoredProcedure("nextForState");
+            cosmosDbApi.uploadStoredProcedure("lease");
+        }
     }
 }
